+++ Sandbox/Sandbox/Controllers/Passkey/PasskeyNonDiscoverableController.swift
@@ -18,10 +18,9 @@ class PasskeyNonDiscoverableController: UIViewController {
 
     private func login(display mode: Mode) async {
         print("PasskeyNonDiscoverableController.login(display:\(mode))")
-        guard let window = view.window else { fatalError("The view was not in the app's view hierarchy!") }
         guard let username = username.text, !username.isEmpty else { return }
 
-        let request = NativeLoginRequest(anchor: window, origin: "PasskeyNonDiscoverableController.login")
+        let request = NativeLoginRequest(presenting: Presentation(from: self), origin: "PasskeyNonDiscoverableController.login")
         await handleAuthToken {
             try await AppDelegate.reachfive().login(withNonDiscoverableUsername: .Unspecified(username), forRequest: request, usingModalAuthorizationFor: [.Passkey], display: mode)
         }